feat: add profileData.js
